Move purge functions out of event persistence
